+++ src/zope/configuration/config.py
@@ -27,6 +27,7 @@ from zope.schema import URI
 from zope.schema import ValidationError
 
 from zope.configuration.exceptions import ConfigurationError
+from zope.configuration.exceptions import ConfigurationWrapperError
 from zope.configuration.interfaces import IConfigurationContext
 from zope.configuration.interfaces import IGroupingContext
 from zope.configuration.fields import GlobalInterface
@@ -41,7 +42,6 @@ __all__ = [
     'ConfigurationContext',
     'ConfigurationAdapterRegistry',
     'ConfigurationMachine',
-    'ConfigurationExecutionError',
     'IStackItem',
     'SimpleStackItem',
     'RootStackItem',
@@ -74,10 +74,14 @@ testns = 'http://namespaces.zope.org/test'
 
 
 class ConfigurationContext(object):
-    """Mix-in that implements IConfigurationContext
+    """
+    Mix-in for implementing.
+    :class:`zope.configuration.interfaces.IConfigurationContext`.
 
-    Subclasses provide a ``package`` attribute and a ``basepath``
-    attribute.  If the base path is not None, relative paths are
+    Note that this class itself does not actually declare that it
+    implements that interface; the subclass must do that. In addition,
+    subclasses must provide a ``package`` attribute and a ``basepath``
+    attribute. If the base path is not None, relative paths are
     converted to absolute paths using the the base path. If the
     package is not none, relative imports are performed relative to
     the package.
@@ -91,33 +95,37 @@ class ConfigurationContext(object):
     attribute.
 
     The include path is appended to each action and is used when
-    resolving conflicts among actions.  Normally, only the a
+    resolving conflicts among actions. Normally, only the a
     ConfigurationMachine provides the actions attribute. Decorators
     simply use the actions of the context they decorate. The
-    ``includepath`` attribute is a tuple of names.  Each name is
+    ``includepath`` attribute is a tuple of names. Each name is
     typically the name of an included configuration file.
 
     The ``info`` attribute contains descriptive information helpful
-    when reporting errors.  If not set, it defaults to an empty string.
+    when reporting errors. If not set, it defaults to an empty string.
 
-    The actions attribute is a sequence of dictionaries where each dictionary
-    has the following keys:
+    The actions attribute is a sequence of dictionaries where each
+    dictionary has the following keys:
 
-      - ``discriminator``, a value that identifies the action. Two actions
-        that have the same (non None) discriminator conflict.
+        - ``discriminator``, a value that identifies the action. Two
+          actions that have the same (non None) discriminator
+          conflict.
 
-      - ``callable``, an object that is called to execute the action,
+        - ``callable``, an object that is called to execute the
+          action,
 
-      - ``args``, positional arguments for the action
+        - ``args``, positional arguments for the action
 
-      - ``kw``, keyword arguments for the action
+        - ``kw``, keyword arguments for the action
 
-      - ``includepath``, a tuple of include file names (defaults to ())
+        - ``includepath``, a tuple of include file names (defaults to
+          ())
 
-      - ``info``, an object that has descriptive information about
-        the action (defaults to '')
-
+        - ``info``, an object that has descriptive information about
+          the action (defaults to '')
     """
+
+    # pylint:disable=no-member
 
     def __init__(self):
         super(ConfigurationContext, self).__init__()
@@ -666,8 +674,8 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
     info = ''
 
     #: These `Exception` subclasses are allowed to be raised from `execute_actions`
-    #: without being re-wrapped into a `ConfigurationExecutionError`. (`BaseException`
-    #: instances are never wrapped.)
+    #: without being re-wrapped into a `~.ConfigurationError`. (`BaseException`
+    #: and other `~.ConfigurationError` instances are never wrapped.)
     #:
     #: Users of instances of this class may modify this before calling `execute_actions`
     #: if they need to propagate specific exceptions.
@@ -727,9 +735,8 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
             [('f', (1,), {}), ('f', (2,), {})]
 
         If the action raises an error, we convert it to a
-        ConfigurationExecutionError.
+        `~.ConfigurationError`.
 
-            >>> from zope.configuration.config import ConfigurationExecutionError
             >>> output = []
             >>> def bad():
             ...    bad.xxx
@@ -739,22 +746,35 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
             ...   (2, f, (2,)),
             ...   (3, bad, (), {}, (), 'oops')
             ...   ]
-            >>> try:
-            ...    v = context.execute_actions()
-            ... except ConfigurationExecutionError as e:
-            ...    v = e
-            >>> lines = str(v).splitlines()
-            >>> 'AttributeError' in lines[0]
-            True
-            >>> lines[0].endswith("'function' object has no attribute 'xxx'")
-            True
-            >>> lines[1:]
-            ['  in:', '  oops']
+
+            >>> context.execute_actions()
+            Traceback (most recent call last):
+            ...
+            zope.configuration.config.ConfigurationExecutionError: oops
+                AttributeError: 'function' object has no attribute 'xxx'
 
         Note that actions executed before the error still have an effect:
 
             >>> output
             [('f', (1,), {}), ('f', (2,), {})]
+
+        If the exception was already a `~.ConfigurationError`, it is raised
+        as-is with the action's ``info`` added.
+
+            >>> def bad():
+            ...     raise ConfigurationError("I'm bad")
+            >>> context.actions = [
+            ...   (1, f, (1,)),
+            ...   (1, f, (11,), {}, ('x', )),
+            ...   (2, f, (2,)),
+            ...   (3, bad, (), {}, (), 'oops')
+            ...   ]
+            >>> context.execute_actions()
+            Traceback (most recent call last):
+            ...
+            zope.configuration.exceptions.ConfigurationError: I'm bad
+                oops
+
         """
         pass_through_exceptions = self.pass_through_exceptions
         if testing:
@@ -769,29 +789,23 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
                 info = action['info']
                 try:
                     callable(*args, **kw)
+                except ConfigurationError as ex:
+                    ex.add_details(info)
+                    raise
                 except pass_through_exceptions:
                     raise
                 except Exception:
-                    t, v, tb = sys.exc_info()
-                    try:
-                        reraise(ConfigurationExecutionError(t, v, info),
-                                None, tb)
-                    finally:
-                        del t, v, tb
-
+                    # Wrap it up and raise.
+                    reraise(ConfigurationExecutionError(info, sys.exc_info()[1]),
+                            None, sys.exc_info()[2])
         finally:
             if clear:
                 del self.actions[:]
 
-class ConfigurationExecutionError(ConfigurationError):
+class ConfigurationExecutionError(ConfigurationWrapperError):
     """
     An error occurred during execution of a configuration action
     """
-    def __init__(self, etype, evalue, info):
-        self.etype, self.evalue, self.info = etype, evalue, info
-
-    def __str__(self): # pragma: no cover
-        return "%s: %s\n  in:\n  %s" % (self.etype, self.evalue, self.info)
 
 ##############################################################################
 # Stack items
@@ -1671,15 +1685,16 @@ def toargs(context, schema, data):
             try:
                 args[str(name)] = field.fromUnicode(s)
             except ValidationError as v:
-                reraise(ConfigurationError("Invalid value for", n, str(v)),
+                reraise(ConfigurationError("Invalid value for %r" % (n)).add_details(v),
                         None, sys.exc_info()[2])
         elif field.required:
             # if the default is valid, we can use that:
             default = field.default
             try:
                 field.validate(default)
-            except ValidationError:
-                raise ConfigurationError("Missing parameter:", n)
+            except ValidationError as v:
+                reraise(ConfigurationError("Missing parameter: %r" % (n,)).add_details(v),
+                        None, sys.exc_info()[2])
             args[str(name)] = default
 
     if data:
@@ -1801,9 +1816,10 @@ def resolveConflicts(actions):
 class ConfigurationConflictError(ConfigurationError):
 
     def __init__(self, conflicts):
+        super(ConfigurationConflictError, self).__init__()
         self._conflicts = conflicts
 
-    def __str__(self): #pragma NO COVER
+    def _with_details(self, opening, detail_formatter):
         r = ["Conflicting configuration actions"]
         for discriminator, infos in sorted(self._conflicts.items()):
             r.append("  For: %s" % (discriminator, ))
@@ -1811,7 +1827,8 @@ class ConfigurationConflictError(ConfigurationError):
                 for line in text_type(info).rstrip().split(u'\n'):
                     r.append(u"    " + line)
 
-        return "\n".join(r)
+        opening = "\n".join(r)
+        return super(ConfigurationConflictError, self)._with_details(opening, detail_formatter)
 
 
 ##############################################################################